This PR was merged into the 1.3 branch.

Discussion
----------

Fixes #6.

Commits
-------

13c610bb60d4bdbd62442b9fca1a7efc262822fe Set up PHPStan level 1
3f48acc6c49590e81d7d43f46c06c1c64fe871cd Add PHPStan to Travis CI
20ff047d270affda333bc1875b75048e460da3a5 Improve code to match PHPStan's level 2
5c620f74e539e3578f05a1ca41fa3b7506b3be61 Improve code to match PHPStan's level 3
77eddbc913ab98da7dee4c6a56627945c5f8cbed Improve code to match PHPStan's level 4
bfb9405ed54d758533db308a6b6ef98eb3e27f37 Improve code to match PHPStan's level 5
75b7d4cfe737ecb8ba3dd0a28ebda7d10f6ff813 Improve code to match PHPStan's level 6
f572886024fd5b41ded4260e1bb7cc6bfc13da5f Improve code to match PHPStan's level "max"

+++ tests/Translation/TranslatorTest.php
@@ -170,7 +170,7 @@ final class TranslatorTest extends TestCase
         $translator = $this->createTranslator($locale);
 
         $catalogue = new MessageCatalogue($locale);
-        $themeDelimiter = strrpos($locale, '@');
+        $themeDelimiter = (int) strrpos($locale, '@');
 
         $catalogue->addFallbackCatalogue(new MessageCatalogue(substr($locale, 0, $themeDelimiter)));
